Step 2 for clean code
