+++ ail_typo_squatting/generator/omission.py
@@ -1,7 +1,14 @@
 # Import all the utils
 from .utils.generator_functions import *
 
+"""
 
+   Original Domain        Typosquatted Domain
+  +----------------+     +--------------------+
+  |    circl.lu    |     |      cirl.lu       |
+  +----------------+     +--------------------+
+
+"""
 
 # Omission
 def omission(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
