+++ ail_typo_squatting/generator/repetition.py
@@ -1,6 +1,14 @@
 # Import all the utils
 from .utils.generator_functions import *
 
+"""
+
+   Original Domain        Typosquatted Domain
+  +----------------+     +----------------------+
+  |    circl.lu    |     |       circll.lu      |
+  +----------------+     +----------------------+
+
+"""
 
 # Repetition
 def repetition(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
